Moved augmenting_db_json to pepseq.
